+++ interfaces/CalendarItem.php
@@ -90,4 +90,21 @@ interface CalendarItem
      * @return string
      */
     public function getIcon();
+
+    /**
+     * @return string
+     */
+    public function getUid();
+
+    /**
+     * @return boolean
+     */
+    public function isExportable();
+
+    /**
+     * @return string
+     */
+    public function getRRule();
+
+    public function getExdate();
 }