+++ views/global/index.php
@@ -5,16 +5,14 @@ use humhub\modules\calendar\widgets\FullCalendar;
 use yii\helpers\Url;
 
 /* @var $this \humhub\components\View */
-/* @var $canConfigure boolean */
-/* @var $configureUrl string */
+/* @var $selectors array */
+/* @var $filters array */
 ?>
 <div class="container">
     <div class="panel panel-default">
         <div class="panel-body" style="background-color:<?= $this->theme->variable('background-color-secondary') ?>">
             <?= CalendarFilterBar::widget([
                     'selectors' => $selectors,
-                    'canConfigure' => $canConfigure,
-                    'configUrl' => $configureUrl,
                     'filters' => $filters
             ])?>
         </div>
@@ -23,15 +21,13 @@ use yii\helpers\Url;
 
                 <div class="panel panel-default" style="margin-bottom:0px">
                     <div class="panel-body">
-                        <?=
-                        FullCalendar::widget([
-                            'canWrite' => true,
+                        <?= FullCalendar::widget([
+                            'canWrite' => !Yii::$app->user->isGuest,
                             'selectors' => $selectors,
                             'filters' => $filters,
                             'loadUrl' => Url::to(['load-ajax']),
                             'editUrl' => $editUrl,
-                        ]);
-                        ?>
+                        ]); ?>
                     </div>
                 </div>